Prove a plain state is plain, and split the form-filling suite
The "adds a plain state" scenario only checked the name appeared, so a form
that rendered a job or deposit ticked by default would have added a state
carrying a marker and the story would still have passed. It now reads the
markers on that state's own row and expects none. Proved by defaulting the
starting-state box to ticked: the new step is the one that fails.

`markersOnRow` replaces the badge check it grew out of, and the "no longer
marks" reader is a plain lookup in that list.

The form-filling suite had reached 460 lines. Its three groups are now three
files named for what each is about, all well under the limit: checking a whole
form's values, filling one in and sending it, and taking a thing down.

The split surfaced two readers sharing a signature, so `ReadAboutOneThing`
takes what it hands back, and both say `ReadAboutOneThing<string[]>` rather
than spelling the same contract out twice.
